backend/cms/cms/serializers/tag_serializer.py
from rest_framework import serializers
from cms.models.tag_model import Tag

class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = ['id', 'name', 'slug']
