backend/cms/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
bleach==5.0.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-appconf==1.0.6
django-cors-headers==4.4.0
django-decouple==2.1
django-environ==0.11.2
django-imagekit==5.0.0
django-meta==2.4.2
django-prometheus==2.3.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
grafana==0.0.1
gunicorn==22.0.0
inflection==0.5.1
kombu==5.4.0
packaging==24.1
pilkit==3.0
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
rest-framework-simplejwt==0.0.2
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
